ref(seer): Move get_supported_scm_providers to seer_setup
Move the function from constants.py to seer_setup.py where other
feature-flag-gated Seer capability checks live. constants.py should
only hold static values.
