+++ src/sentry/seer/constants.py
@@ -1,11 +1,6 @@
-from __future__ import annotations
-
-from typing import TYPE_CHECKING, Literal
+from typing import Literal
 
 from sentry.integrations.types import IntegrationProviderSlug
-
-if TYPE_CHECKING:
-    from sentry.models.organization import Organization
 
 # Type for Seer-supported SCM provider strings
 SeerSCMProvider = Literal[
@@ -29,12 +24,3 @@ SEER_GITLAB_SCM_PROVIDERS = [
     "integrations:gitlab",
     IntegrationProviderSlug.GITLAB.value,
 ]
-
-
-def get_supported_scm_providers(organization: Organization | None = None) -> list[str]:
-    from sentry import features
-
-    providers = list(SEER_SUPPORTED_SCM_PROVIDERS)
-    if organization is not None and features.has("organizations:seer-gitlab-support", organization):
-        providers.extend(SEER_GITLAB_SCM_PROVIDERS)
-    return providers
