+++ src/sentry/seer/seer_setup.py
@@ -3,8 +3,16 @@ from django.contrib.auth.models import AnonymousUser
 from sentry import features
 from sentry.models.organization import Organization
 from sentry.organizations.services.organization.model import RpcOrganization
+from sentry.seer.constants import SEER_GITLAB_SCM_PROVIDERS, SEER_SUPPORTED_SCM_PROVIDERS
 from sentry.users.models.user import User
 from sentry.users.services.user.model import RpcUser
+
+
+def get_supported_scm_providers(organization: Organization | None = None) -> list[str]:
+    providers = list(SEER_SUPPORTED_SCM_PROVIDERS)
+    if organization is not None and features.has("organizations:seer-gitlab-support", organization):
+        providers.extend(SEER_GITLAB_SCM_PROVIDERS)
+    return providers
 
 
 def has_seer_access(
